Move routes content to pages directory
