+++ src/pages/Contact.js
@@ -1,4 +1,5 @@
 import React from "react";
+import '../styles/Contact.css';
 import { Container, Row, Col, Form, Button } from "react-bootstrap";
 import { ImTwitter, ImMail, ImGithub } from "react-icons/im";
 import { FaGitter } from "react-icons/fa";
