+++ src/pages/Journal.js
